Add Conversion Logic

+++ app/src/main/java/com/example/testproject/presentation/rate/model/BaseRateCurrencyUIContent.kt
@@ -1,6 +1,8 @@
 package com.example.testproject.presentation.rate.model
 
+import java.math.BigDecimal
+
 data class BaseRateCurrencyUIContent(
     val currency: RateCurrencyUIContent,
-    val amount: Double
+    val amount: BigDecimal
 )
